+++ src/components/policies/PoliciesDialog.tsx
@@ -20,6 +20,7 @@
 
 import { Button, Dialog, DialogActions, DialogContent, DialogHeader } from 'cx-portal-shared-components';
 import { useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
 
 import { setPageLoading } from '../../features/app/slice';
 import { setSnackbarMessage } from '../../features/notifiication/slice';
@@ -69,6 +70,7 @@ export default function PoliciesDialog() {
   const { currentUploadData, selectedFiles } = useAppSelector(state => state.providerSlice);
   const { selectedSubmodel } = useAppSelector(state => state.submodelSlice);
   const [showError, setshowError] = useState(false);
+  const { t } = useTranslation();
 
   const dispatch = useAppDispatch();
 
@@ -86,6 +88,7 @@ export default function PoliciesDialog() {
     dispatch(setPageLoading(false));
     dispatch(clearRows());
     dispatch(handleDialogClose());
+    dispatch(removeSelectedFiles());
   };
 
   const processingReport = (r: { data: ProcessReport }, processId: string) => {
@@ -108,25 +111,24 @@ export default function PoliciesDialog() {
     } else {
       clearUpload();
       dispatch(setUploadData(defaultUploadData));
-      dispatch(removeSelectedFiles());
       if (r?.data?.status === Status.completed && r?.data?.numberOfFailedItems === 0) {
         dispatch(
           setSnackbarMessage({
-            message: 'Upload completed!',
+            message: t('alerts.uploadSuccess'),
             type: 'success',
           }),
         );
       } else if (r?.data?.status === Status.completed && r?.data?.numberOfFailedItems > 0) {
         dispatch(
           setSnackbarMessage({
-            message: 'Upload completed with warnings!',
+            message: t('alerts.uploadWarning'),
             type: 'warning',
           }),
         );
       } else {
         dispatch(
           setSnackbarMessage({
-            message: 'Upload failed!',
+            message: t('alerts.uploadError'),
             type: 'error',
           }),
         );
@@ -146,6 +148,12 @@ export default function PoliciesDialog() {
           if (error.response.status === 404) {
             processingReportFirstCall(processId);
           } else {
+            dispatch(
+              setSnackbarMessage({
+                message: t('alerts.uploadError'),
+                type: 'error',
+              }),
+            );
             clearUpload();
           }
         });
@@ -187,6 +195,12 @@ export default function PoliciesDialog() {
       // first call
       processingReportFirstCall(response.data);
     } catch (error) {
+      dispatch(
+        setSnackbarMessage({
+          message: t('alerts.uploadError'),
+          type: 'error',
+        }),
+      );
       dispatch(setUploadData({ ...currentUploadData, status: Status.failed }));
       clearUpload();
     }
@@ -196,16 +210,22 @@ export default function PoliciesDialog() {
     const formData = new FormData();
     formData.append('file', selectedFiles[0]);
     formData.append('meta_data', JSON.stringify(payload));
-    formData.append('submodel', selectedSubmodel);
+    formData.append('submodel', selectedSubmodel.value);
 
     try {
       dispatch(setPageLoading(true));
-      const resp = await ProviderService.getInstance().uploadData(selectedSubmodel, formData);
+      const resp = await ProviderService.getInstance().uploadData(selectedSubmodel.value, formData);
       const processId = resp.data;
       // first call
       processingReportFirstCall(processId);
     } catch (error) {
       dispatch(setUploadData({ ...currentUploadData, status: Status.failed }));
+      dispatch(
+        setSnackbarMessage({
+          message: t('alerts.uploadError'),
+          type: 'error',
+        }),
+      );
       clearUpload();
     }
   };
@@ -227,17 +247,21 @@ export default function PoliciesDialog() {
   return (
     // Dialog width change is not available currently in cx-shared-components library
     <Dialog open={openDialog}>
-      <DialogHeader closeWithIcon onCloseWithIcon={() => dispatch(handleDialogClose())} title="Policies" />
+      <DialogHeader
+        closeWithIcon
+        onCloseWithIcon={() => dispatch(handleDialogClose())}
+        title={t('content.policies.title')}
+      />
       <DialogContent>
         <AccessPolicy />
         <UsagePolicy />
       </DialogContent>
       <DialogActions>
         <Button variant="contained" sx={{ mr: 2 }} onClick={() => dispatch(handleDialogClose())}>
-          Close
+          {t('button.close')}
         </Button>
         <Button variant="contained" onClick={handleSubmitData}>
-          Submit
+          {t('button.submit')}
         </Button>
       </DialogActions>
     </Dialog>