+++ src/features/app/slice.ts
@@ -21,7 +21,7 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
 import { IUser } from '../../models/User';
-import { fetchUserPermissions } from './actions';
+import { fetchUseCases, fetchUserPermissions } from './actions';
 import { IAppSlice } from './types';
 
 const initialState: IAppSlice = {
@@ -30,12 +30,17 @@ const initialState: IAppSlice = {
     userName: '',
     name: '',
     email: '',
+    roles: [],
     company: '',
+    bpn: '',
     tenant: '',
     token: '',
     parsedToken: {},
   },
   permissions: [],
+  useCases: [],
+  sidebarExpanded: true,
+  selectedUseCases: [],
 };
 export const appSlice = createSlice({
   name: 'appSlice',
@@ -47,6 +52,12 @@ export const appSlice = createSlice({
     setLoggedInUser: (state, action: PayloadAction<IUser>) => {
       state.loggedInUser = action.payload;
     },
+    setSidebarExpanded: state => {
+      state.sidebarExpanded = !state.sidebarExpanded;
+    },
+    setSelectedUseCases: (state, action: PayloadAction<string[]>) => {
+      state.selectedUseCases = action.payload;
+    },
   },
   extraReducers: builder => {
     builder.addCase(fetchUserPermissions.pending, state => {
@@ -56,8 +67,15 @@ export const appSlice = createSlice({
       state.permissions = action.payload;
       state.pageLoading = false;
     });
+    builder.addCase(fetchUseCases.pending, state => {
+      state.pageLoading = true;
+    });
+    builder.addCase(fetchUseCases.fulfilled, (state, action) => {
+      state.useCases = action.payload;
+      state.pageLoading = false;
+    });
   },
 });
 
-export const { setPageLoading, setLoggedInUser } = appSlice.actions;
+export const { setPageLoading, setLoggedInUser, setSelectedUseCases, setSidebarExpanded } = appSlice.actions;
 export default appSlice.reducer;