+++ src/features/app/types.ts
@@ -23,4 +23,12 @@ export interface IAppSlice {
   pageLoading: boolean;
   loggedInUser: IUser;
   permissions: string[];
+  useCases: IUseCase[];
+  selectedUseCases: string[];
+  sidebarExpanded: boolean;
+}
+
+export interface IUseCase {
+  id: string;
+  title: string;
 }