+++ src/index.tsx
@@ -27,11 +27,13 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
 import App from './App';
+import I18nService from './services/i18nService';
 import UserService from './services/UserService';
 import { store } from './store/store';
 import { clearConsoles } from './utils/utils';
 
 clearConsoles();
+I18nService.init();
 
 UserService.initKeycloak(user => {
   ReactDOM.render(