+++ src/models/Sidebar.tsx
@@ -32,54 +32,48 @@ export const icons = {
   HomeIcon,
 };
 
-export interface IntMenuChildren {
+export interface IntMenuItem {
   text: string;
-  menuIcon?: keyof typeof icons;
   to?: string;
+  menuIcon?: keyof typeof icons;
   dataId?: string;
+  isHeading?: boolean;
   permissions?: string[];
 }
 
-export interface IntMenuItem {
-  text: string;
-  isHeading?: boolean;
-  isExpanded?: boolean;
-  childrens?: IntMenuChildren[];
-}
-
-export interface IntMenuItemProps {
-  item: IntMenuChildren;
-  isExpanded: boolean;
+export interface IntMenu extends IntMenuItem {
+  childrens?: IntMenuItem[];
 }
 
 // menu items
-export const MenuItems: IntMenuItem[] = [
+export const MenuItems: IntMenu[] = [
   {
-    text: 'Provider',
+    text: 'pages.home',
+    menuIcon: 'HomeIcon',
+    to: '/',
+    dataId: 'homePage',
+    isHeading: false,
+  },
+  {
+    text: 'pages.provider',
     isHeading: true,
     childrens: [
-      // {
-      //   text: 'Home',
-      //   menuIcon: 'HomeIcon',
-      //   to: '/',
-      //   dataId: 'homePage',
-      // },
       {
-        text: 'Create Data',
+        text: 'pages.createData',
         menuIcon: 'AddCircleIcon',
         to: '/create-data',
         dataId: 'uploadFileMenu',
         permissions: ['provider_create_contract_offer'],
       },
       {
-        text: 'Upload History',
+        text: 'pages.uploadHistory',
         menuIcon: 'HistoryIcon',
         to: '/upload-history',
         dataId: 'uploadHistoryMenu',
         permissions: ['provider_view_history'],
       },
       {
-        text: 'Help',
+        text: 'pages.help',
         menuIcon: 'HelpIcon',
         to: '/help',
         dataId: 'helpMenu',
@@ -87,11 +81,11 @@ export const MenuItems: IntMenuItem[] = [
     ],
   },
   {
-    text: 'Consumer',
+    text: 'pages.consumer',
     isHeading: true,
     childrens: [
       {
-        text: 'Consume Data',
+        text: 'pages.consumeData',
         menuIcon: 'ManageSearchIcon',
         to: '/consume-data',
         dataId: 'uploadFileMenu',
@@ -103,7 +97,7 @@ export const MenuItems: IntMenuItem[] = [
         ],
       },
       {
-        text: 'Contract History',
+        text: 'pages.contractHistory',
         menuIcon: 'HistoryIcon',
         to: '/contract-history',
         dataId: 'uploadHistoryMenu',