+++ src/models/User.ts
@@ -25,6 +25,8 @@ export interface IUser {
   name: string;
   email: string;
   company: string;
+  roles: string[];
+  bpn: string;
   tenant: string;
   token: string;
   parsedToken: KeycloakTokenParsed;