+++ src/pages/ContractHistory.tsx
@@ -26,6 +26,7 @@ import { Box, Chip, Grid, LinearProgress, Stack, Typography } from '@mui/materia
 import { DataGrid, GridRenderCellParams, GridToolbar, GridValueGetterParams } from '@mui/x-data-grid';
 import { Button } from 'cx-portal-shared-components';
 import React, { useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
 
 import Permissions from '../components/Permissions';
 import { handleBlankCellValues, MAX_CONTRACTS_AGREEMENTS } from '../helpers/ConsumerOfferHelper';
@@ -39,6 +40,7 @@ const ContractHistory: React.FC = () => {
   const [pageSize, setPageSize] = useState<number>(10);
   const { contractAgreements, isContractAgreementsLoading } = useAppSelector(state => state.consumerSlice);
   const dispatch = useAppDispatch();
+  const { t } = useTranslation();
 
   const renderContractAgreementStatus = (params: GridRenderCellParams) => {
     switch (params.value) {
@@ -80,50 +82,38 @@ const ContractHistory: React.FC = () => {
     {
       field: 'contractAgreementId',
       flex: 1,
-      editable: false,
-      headerName: 'Contract Agreement ID',
-      renderHeader: () => <strong>Contract Agreement ID</strong>,
+      headerName: t('content.contractHistory.columns.contractAgreementId'),
       valueGetter: (params: GridValueGetterParams) => handleBlankCellValues(params.row.contractAgreementId),
     },
     {
       field: 'contractAgreementInfo.assetId',
       flex: 1,
-      editable: false,
-      headerName: 'Asset ID',
-      renderHeader: () => <strong>Asset ID</strong>,
+      headerName: t('content.contractHistory.columns.assetId'),
       valueGetter: (params: GridValueGetterParams) =>
         params.row.contractAgreementInfo ? params.row.contractAgreementInfo?.assetId : '-',
     },
     {
       field: 'counterPartyAddress',
       flex: 1,
-      editable: false,
-      headerName: 'Counter Party Address',
-      renderHeader: () => <strong>Counter Party Address</strong>,
+      headerName: t('content.contractHistory.columns.counterPartyAddress'),
       valueGetter: (params: GridValueGetterParams) => handleBlankCellValues(params.row.counterPartyAddress),
     },
     {
       field: 'title',
       flex: 1,
-      editable: false,
-      headerName: 'Title',
-      renderHeader: () => <strong>Title</strong>,
+      headerName: t('content.contractHistory.columns.title'),
       valueGetter: (params: GridValueGetterParams) => handleBlankCellValues(params.row.title),
     },
     {
       field: 'organizationName',
       flex: 1,
-      editable: false,
-      headerName: 'Organization',
-      renderHeader: () => <strong>Organization</strong>,
+      headerName: t('content.contractHistory.columns.organizationName'),
       valueGetter: (params: GridValueGetterParams) => handleBlankCellValues(params.row.organizationName),
     },
     {
       field: 'contractAgreementInfo.contractSigningDate',
       flex: 1,
-      editable: false,
-      headerName: 'Signing Date',
-      renderHeader: () => <strong>Signing Date</strong>,
+      headerName: t('content.contractHistory.columns.contractSigningDate'),
       valueGetter: (params: GridValueGetterParams) =>
         params.row.contractAgreementInfo?.contractSigningDate
           ? convertEpochToDate(params.row.contractAgreementInfo.contractSigningDate)
@@ -132,9 +122,7 @@ const ContractHistory: React.FC = () => {
     {
       field: 'contractAgreementInfo.contractEndDate',
       flex: 1,
-      editable: false,
-      headerName: 'End Date',
-      renderHeader: () => <strong>End Date</strong>,
+      headerName: t('content.contractHistory.columns.contractEndDate'),
       valueGetter: (params: GridValueGetterParams) =>
         params.row.contractAgreementInfo?.contractEndDate
           ? convertEpochToDate(params.row.contractAgreementInfo.contractEndDate)
@@ -143,9 +131,7 @@ const ContractHistory: React.FC = () => {
     {
       field: 'state',
       flex: 1,
-      editable: false,
-      headerName: 'Status',
-      renderHeader: () => <strong>Status</strong>,
+      headerName: t('content.contractHistory.columns.state'),
       renderCell: renderContractAgreementStatus,
     },
   ];
@@ -176,15 +162,15 @@ const ContractHistory: React.FC = () => {
 
   return (
     <Box sx={{ flex: 1, p: 4 }}>
-      <Permissions values={['consumer_view_contract_agreement']}>
+      <Permissions values={['consumer_view_contract_agreement']} fullPage={true}>
         <Grid container spacing={2} alignItems="center">
           <Grid item xs={6} my={4}>
-            <Typography variant="h4">Contract Agreements History</Typography>
+            <Typography variant="h4">{t('content.contractHistory.title')}</Typography>
           </Grid>
           <Grid item xs={6} my={4} textAlign={'right'}>
             <Button size="small" variant="contained" onClick={() => fetchContractAgreements()}>
               <Refresh />
-              &nbsp; Refresh
+              &nbsp; {t('button.refresh')}
             </Button>
           </Grid>
           <Grid item xs={12}>
@@ -205,12 +191,12 @@ const ContractHistory: React.FC = () => {
                   LoadingOverlay: LinearProgress,
                   NoRowsOverlay: () => (
                     <Stack height="100%" alignItems="center" justifyContent="center">
-                      No Contract agreements!
+                      {t('content.common.noData')}
                     </Stack>
                   ),
                   NoResultsOverlay: () => (
                     <Stack height="100%" alignItems="center" justifyContent="center">
-                      Contract agreements not found!
+                      {t('content.common.noResults')}
                     </Stack>
                   ),
                 }}