+++ src/pages/UploadHistory.tsx
@@ -22,18 +22,20 @@
 import { Refresh } from '@mui/icons-material';
 import { Box, Grid } from '@mui/material';
 import { Button, Typography } from 'cx-portal-shared-components';
-import React, { useCallback, useEffect, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
 
 import Permissions from '../components/Permissions';
 import StickyHeadTable from '../components/StickyHeadTable';
 import { ProcessReport } from '../models/ProcessReport';
 import ProviderService from '../services/ProviderService';
 
-export const UploadHistory: React.FC = () => {
+export default function UploadHistory() {
   const [tableData, setTableData] = useState<ProcessReport[]>([]);
   const [rowsPerPage, setRowsPerPage] = useState<number>(10);
   const [page, setPage] = useState<number>(0);
   const [totalElements, setTotalElements] = useState<number>(0);
+  const { t } = useTranslation();
 
   const refreshTable = useCallback(async () => {
     try {
@@ -56,16 +58,16 @@ export const UploadHistory: React.FC = () => {
   }, [page, rowsPerPage, refreshTable]);
 
   return (
-    <Permissions values={['provider_view_history']}>
+    <Permissions values={['provider_view_history']} fullPage={true}>
       <Box sx={{ flex: 1, p: 4 }}>
         <Grid container spacing={2} alignItems="center">
           <Grid item xs={6}>
-            <Typography variant="h4">Upload History</Typography>
+            <Typography variant="h4">{t('pages.uploadHistory')}</Typography>
           </Grid>
           <Grid item xs={6} textAlign="right">
             <Button size="small" variant="contained" onClick={() => refreshTable()}>
               <Refresh />
-              &nbsp; Refresh
+              &nbsp; {t('button.refresh')}
             </Button>
           </Grid>
         </Grid>
@@ -83,6 +85,4 @@ export const UploadHistory: React.FC = () => {
       </Box>
     </Permissions>
   );
-};
-
-export default UploadHistory;
+}