+++ src/services/ProviderService.ts
@@ -32,10 +32,12 @@ class ProviderService extends HttpService {
     return this.classInstance;
   }
 
-  public async getSubmodelList() {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  public async getSubmodelList(params?: any) {
+    const usecaseParam = params.length ? `?usecases=${params}` : '';
     const res = await this.instance({
       method: 'GET',
-      url: '/submodels',
+      url: `/submodels${usecaseParam}`,
     });
     return res?.data;
   }