+++ src/services/appService.ts
@@ -41,6 +41,30 @@ class AppService extends HttpService {
     });
     return res;
   }
+
+  public async downloadCSV(submodel: string, type: string) {
+    const res = await this.instance({
+      method: 'GET',
+      url: `/submodels/csvfile/${submodel}?type=${type}`,
+    });
+    return res;
+  }
+
+  public async downloadHistory(submodel: string, processId: string) {
+    const res = await this.instance({
+      method: 'GET',
+      url: `/${submodel}/download/${processId}/csv`,
+    });
+    return res;
+  }
+
+  public async getUseCases() {
+    const res = await this.instance({
+      method: 'GET',
+      url: '/usecases',
+    });
+    return res;
+  }
 }
 
 export default AppService;